rustc: move infer::coercion to rustc_typeck.
